timeaware deletions

+++ dr/views.py
@@ -1,6 +1,8 @@
 from dr.models import Room, Reservation, User
 from django.core.mail import send_mail
 from django.conf import settings
+import pytz
+from django.utils import timezone
 
 from django.shortcuts import render, redirect
 from django.http import HttpResponse, Http404, HttpResponseRedirect
@@ -13,7 +15,7 @@ from .forms import UserCreateForm, ReservationForm
 from .models import User
 from django.contrib.auth.models import User as muser
 from django.contrib.auth.hashers import make_password
-from datetime import datetime
+from datetime import datetime, timedelta
 from.google_calendar import Calendar
 
 def index(request):
@@ -104,12 +106,17 @@ def reservations(request):
         if request.GET.get('delete'):
             id = request.GET.get('id')
             instance = Reservation.objects.filter(id=id)
-            print("---- deleting ----",id)
-            if instance[0].googleId:
-                print("---- deletin - googleId ----", instance[0].googleId)
-                calendar = Calendar()
-                calendar.deleteEvent(instance[0].googleId)
-            instance.delete()  
+            
+            now = timezone.now()
+            start = instance[0].start_reservation
+
+            if now + timedelta(hours=24) <= start: 
+             if instance[0].googleId:
+                 print("---- deletin - googleId ----", instance[0].googleId)
+                 calendar = Calendar()
+                 calendar.deleteEvent(instance[0].googleId)
+             print("---- deleting ----",id)
+             instance.delete()  
         return render(request, 'reservations.html', context=context)
     else:
         return HttpResponseRedirect('/dr/login')
